test(file): add TestFilterLines
